+++ src/TradingTerminal.Login/Forms/BinanceLoginForm.xaml.cs
@@ -1,0 +1,11 @@
+using System.Windows.Controls;
+
+namespace TradingTerminal.App.Login.Forms;
+
+public partial class BinanceLoginForm : UserControl
+{
+    public BinanceLoginForm()
+    {
+        InitializeComponent();
+    }
+}